requirements.txt
markitdown
openai
python-dotenv